backend/requirements.txt
langchain
qdrant-client
sentence-transformers
streamlit
python-dotenv
requests